refactor: change project architecture to domain driven
